# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gato API",
    "version": "1.0.0",
    "description": "Translation, vocabulary, and spaced-repetition API powering the Gato Chrome extension and web dashboard."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "WordOfTheDay": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string"
          },
          "translation": {
            "type": "string"
          },
          "sourceLanguage": {
            "type": "string"
          },
          "targetLanguage": {
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "exampleSentence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "type": "string"
          }
        },
        "required": [
          "word",
          "translation",
          "sourceLanguage",
          "targetLanguage",
          "rationale",
          "exampleSentence",
          "date"
        ],
        "additionalProperties": false
      },
      "RelatedSuggestion": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string"
          },
          "translation": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "exampleSentence": {
            "type": "string"
          }
        },
        "required": [
          "word",
          "translation",
          "rationale",
          "exampleSentence"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/suggestions/word-of-the-day": {
      "get": {
        "summary": "Today's suggested vocabulary word",
        "tags": [
          "suggestions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestion": {
                      "$ref": "#/components/schemas/WordOfTheDay"
                    }
                  },
                  "required": [
                    "suggestion"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/suggestions/related-to-vocab": {
      "get": {
        "summary": "Vocabulary suggestions related to saved concepts",
        "tags": [
          "suggestions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RelatedSuggestion"
                      }
                    }
                  },
                  "required": [
                    "suggestions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.gato.app",
      "description": "Production"
    }
  ],
...
}